Refactored adding serde defaults, removing unused public functionality. Added MS HyperV NetVSC virtual device

+++ workspace/dpdk-virtual-devices/src/net_virtual_devices/BondingSlave.rs
@@ -18,7 +18,7 @@ pub enum BondingSlave
 	ByVirtualDeviceName(NetVirtualDeviceName),
 	
 	/// By ethernet port identifier.
-	ByEthernetPortIdentifier(u32),
+	ByEthernetPortIdentifier(u16),
 }
 
 impl BondingSlave
